+++ source/Energinet.DataHub.MarketParticipant.Application/Commands/Query/User/AssignedActorDto.cs
@@ -12,17 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
 using System.Collections.Generic;
 
-namespace Energinet.DataHub.MarketParticipant.Client.Models
-{
-    public sealed record UserOverviewItemDto(
-            Guid Id,
-            UserStatus Status,
-            string Name,
-            string Email,
-            string? PhoneNumber,
-            DateTimeOffset CreatedDate,
-            IEnumerable<AssignedActorDto> AssignedActors);
-}
+namespace Energinet.DataHub.MarketParticipant.Application.Commands.Query.User;
+
+public sealed record AssignedActorDto(ActorDto Actor, IEnumerable<string> UserRoleNames);