+++ source/Energinet.DataHub.MarketParticipant.Client/source/Energinet.DataHub.MarketParticipant.Client/Models/AssignedActorDto.cs
@@ -12,17 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
 using System.Collections.Generic;
 
 namespace Energinet.DataHub.MarketParticipant.Client.Models
 {
-    public sealed record UserOverviewItemDto(
-            Guid Id,
-            UserStatus Status,
-            string Name,
-            string Email,
-            string? PhoneNumber,
-            DateTimeOffset CreatedDate,
-            IEnumerable<AssignedActorDto> AssignedActors);
+    public sealed record AssignedActorDto(UserOverviewItemActorDto Actor, IEnumerable<string> UserRoleNames);
 }